+++ FmuImporter/FmuImporter/FmuImporter.cs
@@ -279,6 +279,8 @@ public class FmuImporter
       SilKitDataManager.CreateSubscriber(
         configuredVariable.FmuVariableDefinition.Name,
         configuredVariable.TopicName,
+        _fmuImporterConfig.SubscriberInstance,
+        _fmuImporterConfig.Namespace,
         new IntPtr(configuredVariable.FmuVariableDefinition.ValueReference));
     }
 
@@ -288,6 +290,8 @@ public class FmuImporter
       SilKitDataManager.CreateSubscriber(
         configuredStructure.Name,
         configuredStructure.Name,
+        _fmuImporterConfig.SubscriberInstance,
+        _fmuImporterConfig.Namespace,
         new IntPtr(configuredStructure.StructureId));
     }
 
@@ -308,6 +312,8 @@ public class FmuImporter
       SilKitDataManager.CreatePublisher(
         configuredVariable.FmuVariableDefinition.Name,
         configuredVariable.TopicName,
+        _fmuImporterConfig.PublisherInstance,
+        _fmuImporterConfig.Namespace,
         new IntPtr(configuredVariable.FmuVariableDefinition.ValueReference),
         0);
     }
@@ -315,10 +321,11 @@ public class FmuImporter
     // create publishers for output structures
     foreach (var configuredStructure in FmuDataManager.OutputConfiguredStructures.Values)
     {
-      // TODO add label support
       SilKitDataManager.CreatePublisher(
         configuredStructure.Name,
         configuredStructure.Name,
+        _fmuImporterConfig.PublisherInstance,
+        _fmuImporterConfig.Namespace,
         new IntPtr(configuredStructure.StructureId),
         0);
     }
@@ -340,6 +347,8 @@ public class FmuImporter
       SilKitDataManager.CreatePublisher(
         configuredVariable.FmuVariableDefinition.Name,
         configuredVariable.TopicName,
+        _fmuImporterConfig.PublisherInstance,
+        _fmuImporterConfig.Namespace,
         new IntPtr(configuredVariable.FmuVariableDefinition.ValueReference),
         0);
     }